moved hivalc specific functions out to mappings
